Add files via upload
Logistic regression is a popular statistical modeling technique used to predict binary outcomes or classify data into discrete categories. In Python, logistic regression can be implemented using libraries such as pandas, numpy, and scikit-learn.

Logistic regression deals with binary or categorical outcomes. It models the relationship between one or more independent variables (features) and a binary dependent variable (target) using the logistic function, also known as the sigmoid function.

In logistic regression, the goal is to estimate the probability of an event occurring based on the given set of features. The logistic function maps any real-valued number to a value between 0 and 1, which represents the probability of the event occurring.